change: top layer, fallen block, and erosion config for megalith
added a block state provider for the top layer, a chance for the top
layer, int providers for the fallen and floating blocks, and completely
reworked erosion to be based off of a mask and chance, from the bottom
upwards. The mask decreases based on distance from the bottom. Two
dimensionally, the mask is based on the sqrt of euclidean distance from
the center of the megalith. Vertically, decreases based on its
proportion of space of the total allocated erosion space.